+++ src/TechChallenge.Application/Features/Funcionarios/ExcluirFuncionario/ExcluirFuncionarioService.cs
@@ -6,11 +6,16 @@ namespace TechChallenge.Application.Features.Funcionarios.ExcluirFuncionario;
 public class ExcluirFuncionarioService
 {
     private readonly IFuncionarioRepository _funcionarioRepository;
+    private readonly IOrdemServicoRepository _ordemServicoRepository;
     private readonly IValidator<ExcluirFuncionarioCommand> _validator;
 
-    public ExcluirFuncionarioService(IFuncionarioRepository funcionarioRepository, IValidator<ExcluirFuncionarioCommand> validator)
+    public ExcluirFuncionarioService(
+        IFuncionarioRepository funcionarioRepository,
+        IOrdemServicoRepository ordemServicoRepository,
+        IValidator<ExcluirFuncionarioCommand> validator)
     {
         _funcionarioRepository = funcionarioRepository;
+        _ordemServicoRepository = ordemServicoRepository;
         _validator = validator;
     }
 
@@ -21,6 +26,10 @@ public class ExcluirFuncionarioService
         var funcionario = _funcionarioRepository.GetByIdAsync(command.Id).GetAwaiter().GetResult();
         if (funcionario is null)
             throw new KeyNotFoundException($"Funcionário com Id {command.Id} não encontrado.");
+
+        var funcionarioPossuiOrdemServico = _ordemServicoRepository.ExistePorFuncionarioAsync(command.Id).GetAwaiter().GetResult();
+        if (funcionarioPossuiOrdemServico)
+            throw new InvalidOperationException("Não é possível excluir um funcionário associado a uma ordem de serviço.");
 
         _funcionarioRepository.DeleteAsync(funcionario).GetAwaiter().GetResult();
         return true;
