+++ src/TechChallenge.Application/Features/Veiculos/ExcluirVeiculo/ExcluirVeiculoService.cs
@@ -6,11 +6,16 @@ namespace TechChallenge.Application.Features.Veiculos.ExcluirVeiculo;
 public class ExcluirVeiculoService
 {
     private readonly IVeiculoRepository _veiculoRepository;
+    private readonly IOrdemServicoRepository _ordemServicoRepository;
     private readonly IValidator<ExcluirVeiculoCommand> _validator;
 
-    public ExcluirVeiculoService(IVeiculoRepository veiculoRepository, IValidator<ExcluirVeiculoCommand> validator)
+    public ExcluirVeiculoService(
+        IVeiculoRepository veiculoRepository,
+        IOrdemServicoRepository ordemServicoRepository,
+        IValidator<ExcluirVeiculoCommand> validator)
     {
         _veiculoRepository = veiculoRepository;
+        _ordemServicoRepository = ordemServicoRepository;
         _validator = validator;
     }
 
@@ -21,6 +26,10 @@ public class ExcluirVeiculoService
         var veiculo = _veiculoRepository.GetByIdAsync(command.Id).GetAwaiter().GetResult();
         if (veiculo is null)
             throw new KeyNotFoundException($"Veículo com Id {command.Id} não encontrado.");
+
+        var veiculoPossuiOrdemServico = _ordemServicoRepository.ExistePorVeiculoAsync(command.Id).GetAwaiter().GetResult();
+        if (veiculoPossuiOrdemServico)
+            throw new InvalidOperationException("Não é possível excluir um veículo associado a uma ordem de serviço.");
 
         _veiculoRepository.DeleteAsync(veiculo).GetAwaiter().GetResult();
         return true;
